Make showq deal with NULL job names
One of our users discovered that with the DRMAA interface it
is possible to submit jobs with no job name, i.e. a NULL pointer.
This made showq SEGV and crash, which is not good.

This fix checks that job->name is non-NULL before trying to use
it and sets the job name to "(null)" if it is, to match Slurms
behaviour with squeue.