Add more reasons for jobs to be in as defined in slurm.h
I did not find documentation as to what showq consideres 'blocked' jobs and what it
considers 'waiting' jobs. From the existing states I assumed 'blocked' means to
likely not start without user intervention or state changes of other jobs of that
user, except explicit dependencies which are included too. 'waiting' would be
anything else. If this is not correct, please feel free to move the entries
accordingly.

With this in mind, I sorted them such that, e.g., exceeded QOS-limits that are
defined per job are treated as 'blocked', while QOS-limits concerning total
user counts are treated as 'waiting' (for other jobs to finish).